First version of full Backend

+++ backend/PrimeTree/src/main/java/PrimeTree/SQLAdapter.java
@@ -22,6 +22,6 @@ public interface SQLAdapter {
 	 * 
 	 * @return: the listing-object
 	 * */
-	Listing getListingById(int listingId) throws ListingNotFoundException;
+	Listing getListingById(long listingId) throws ListingNotFoundException;
 
 }
